FaceLens-ml/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
face_recognition==1.3.0
python-multipart==0.0.9
Pillow>=10.2.0
numpy>=1.26.4
pydantic==2.6.4
